refactor: extract GovernanceEraseExecutionMixin from governance storage
Move the GovernanceEraseExecution bucket (apply_governance_user_data_delete,
complete_purge_operation_with_audit + the content-purge / clear-user-data
privates) out of the residual GovernanceMixin into a new
GovernanceEraseExecutionMixin, on both the SQLite backend and the backend-neutral
ABC. Behavior-preserving verbatim move: the ERASE-audit-idempotency block, the
status='complete' flip, and the emit-only-on-actual-purge wasPurged lineage guard
are preserved byte-for-byte; cross-bucket collaborators resolve via the composed
MRO. Repoint the _epoch_now completion-retry patch in test_governance_storage to
the new module.
