+++ reflexio/server/services/storage/storage_base/_governance.py
@@ -2,23 +2,12 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 
-from reflexio.models.api_schema.domain.governance import (
-    AuditEvent,
-    PurgeOperation,
-)
-
 
 class GovernanceMixin(ABC):
     """Mixin for backend-neutral governance storage primitives."""
 
     @abstractmethod
     def hide_governance_agent_playbooks_for_rebuild(self, purge_id: str) -> list[int]:
-        raise NotImplementedError
-
-    @abstractmethod
-    def apply_governance_user_data_delete(
-        self, purge_id: str, user_id: str
-    ) -> dict[str, int]:
         raise NotImplementedError
 
     @abstractmethod
@@ -34,10 +23,4 @@ class GovernanceMixin(ABC):
         expanded_terms: str | None,
         tags: list[str] | None,
     ) -> None:
-        raise NotImplementedError
-
-    @abstractmethod
-    def complete_purge_operation_with_audit(
-        self, purge_id: str, audit_event: AuditEvent
-    ) -> PurgeOperation:
         raise NotImplementedError
