+++ reflexio/server/services/storage/storage_base/governance/__init__.py
@@ -1,9 +1,11 @@
 from ._audit import AuditEventStoreMixin
+from ._erase_execution import GovernanceEraseExecutionMixin
 from ._purge import PurgeOperationStoreMixin
 from ._subject_barrier import SubjectBarrierMixin
 
 __all__ = [
     "AuditEventStoreMixin",
+    "GovernanceEraseExecutionMixin",
     "PurgeOperationStoreMixin",
     "SubjectBarrierMixin",
 ]
